+++ src/abb_robot_client/rws2.py
@@ -673,7 +673,7 @@ class RWS2:
             #o.append(RAPIDEventLogEntry(msg_type,code,tstamp,args,title,desc,conseqs,causes,actions))
         return o
     
-    def get_speedratio(self) -> float:
+    def get_speedratio(self) -> int:
         """
         Get the current speed ratio
 
@@ -683,7 +683,7 @@ class RWS2:
         state = res_json["state"][0]
         if not state["_type"] == "pnl-speedratio":
             raise Exception("Invalid speedratio type")
-        return float(state["speedratio"])
+        return int(state["speedratio"])
     
     def set_speedratio(self, speedratio: int):
         """
